# docs/swagger.yaml
basePath: /
definitions:
  elevator.CreateElevatorRequest:
    properties:
      amountStairs:
        type: integer
      heightGain:
        type: integer
      stairs:
        type: boolean
    type: object
  elevator.ElevatorDB:
    properties:
      amountStairs:
        type: integer
      heightGain:
        type: integer
      id:
        type: string
      stairs:
        type: boolean
      time:
        type: integer
      userId:
        type: string
    type: object
  elevator.createElevatorResponse:
    properties:
      id:
        type: string
    type: object
  finance.CreateSpendingRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      saving:
        type: number
      spendingTime:
        type: integer
    type: object
  finance.createSpendingResponse:
    properties:
      id:
        type: string
    type: object
  finance.getInvestmentResponse:
    properties:
      amount:
        type: number
      description:
        type: string
      id:
        type: string
      saving:
        type: number
      spendingTime:
        type: integer
      userId:
        type: string
    type: object
  meditation.CreateMeditationRequest:
    properties:
      endTime:
        type: integer
      meditationTime:
        type: integer
    type: object
  meditation.MeditationDB:
    properties:
      endTime:
        type: integer
      id:
        type: string
      meditationTime:
        type: integer
      userId:
        type: string
    type: object
  meditation.createMeditationResponse:
    properties:
      id:
        type: string
    type: object
  progress.Experience:
    additionalProperties:
      type: number
    type: object
  progress.ExperienceToNewLevel:
    additionalProperties:
      type: number
    type: object
  progress.Response:
    properties:
      experienceToNewLevel:
        $ref: '#/definitions/progress.ExperienceToNewLevel'
      level:
        $ref: '#/definitions/progress.Experience'
    type: object
  settings.CreateSettingsRequest:
    properties:
      elevator:
        $ref: '#/definitions/settings.ElevatorSettings'
      enabledPlugins:
        items:
          $ref: '#/definitions/settings.PluginName'
        type: array
      finance:
        $ref: '#/definitions/settings.FinanceSettings'
      meditation:
        $ref: '#/definitions/settings.MeditationSettings'
    type: object
  settings.ElevatorSettings:
    properties:
      amountNotifications:
        type: integer
      goal:
        type: integer
      notifications:
        type: boolean
      periodNotifications:
        $ref: '#/definitions/settings.NotificationType'
    type: object
  settings.FinanceSettings:
    properties:
      amountNotifications:
        type: integer
      investmentGoal:
        type: integer
      investmentTimeGoal:
        type: integer
      notifications:
        type: boolean
      periodNotifications:
        $ref: '#/definitions/settings.NotificationType'
      strategy:
        $ref: '#/definitions/settings.StrategyType'
      strategyAmount:
        type: integer
    type: object
  settings.MeditationSettings:
    properties:
      amountNotifications:
        type: integer
      meditationTimeGoal:
        type: integer
      notifications:
        type: boolean
      periodNotifications:
        $ref: '#/definitions/settings.NotificationType'
    type: object
  settings.NotificationType:
    enum:
    - Day
    - Month
    - Week
    type: string
    x-enum-varnames:
    - NotificationTypeDay
    - NotificationTypeMonth
    - NotificationTypeWeek
  settings.PluginName:
    enum:
    - finance
    - meditation
    - elevator
    type: string
    x-enum-varnames:
    - PluginNameFinance
    - PluginNameMeditation
    - PluginNameElevator
  settings.StrategyType:
    enum:
    - Round
    - Plus
    - Percent
    type: string
    x-enum-varnames:
    - StrategyTypeRound
    - StrategyTypePlus
    - StrategyTypePercent
  settings.createOnboardingResponse:
    properties:
      id:
        type: string
    type: object
  settings.getPluginSettingResponse:
    properties:
      elevator:
        $ref: '#/definitions/settings.ElevatorSettings'
      enabledPlugins:
        items:
          $ref: '#/definitions/settings.PluginName'
        type: array
      finance:
        $ref: '#/definitions/settings.FinanceSettings'
      id:
        type: string
      meditation:
        $ref: '#/definitions/settings.MeditationSettings'
    type: object
  user.CreateUserRequest:
    properties:
      dateOfBirth:
        type: string
      email:
        type: string
      expoPushToken:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
    type: object
  user.UserDB:
    properties:
      createdAt:
        type: integer
      dateOfBirth:
        type: string
      email:
        type: string
      expoPushToken:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      onboardingDone:
        type: boolean
    type: object
  user.createUserResponse:
    properties:
      id:
        type: string
    type: object
  user.updateUserRequest:
    properties:
      dateOfBirth:
        type: string
      email:
        type: string
      expoPushToken:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      onboardingDone:
        type: boolean
    type: object
info:
  contact:
    name: Wholesome Living
  description: A backend for Wholesome Living written in Golang backend API using
    Fiber and MongoDB
  license:
    name: MIT
  title: Wholesome Living Backend
  version: "0.1"
paths:
  /elevator:
    get:
      description: Fetch one or multiple elevator sessions.
      parameters:
      - description: Elevator ID
        in: query
        name: id
        type: string
      - description: start time
        in: query
        name: startTime
        type: integer
      - description: end time
        in: query
        name: endTime
        type: integer
      - description: duration start time
        in: query
        name: durationStart
        type: integer
      - description: duration end time
        in: query
        name: durationEnd
        type: integer
      - description: Minimum amount of height gained
        in: query
        name: minGain
        type: integer
      - description: Maximum amount of height gained
        in: query
        name: maxGain
        type: integer
      - description: User ID
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/elevator.ElevatorDB'
            type: array
      summary: Get elevator sessions
      tags:
      - elevator
    post:
      consumes:
      - '*/*'
      description: Creates a new elevator.
      parameters:
      - description: Elevator to create
        in: body
        name: elevator
        required: true
        schema:
          $ref: '#/definitions/elevator.CreateElevatorRequest'
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/elevator.createElevatorResponse'
      summary: Create elevator.
      tags:
      - elevator
  /finance:
    get:
      description: Query Investments with the user ID, start time and end time.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: investment ID
        in: query
        name: id
        type: string
      - description: start time
        in: query
        name: startTime
        type: integer
      - description: end time
        in: query
        name: endTime
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.getInvestmentResponse'
      summary: Query Investments with the user ID, start time and end time.
      tags:
      - finance
    post:
      consumes:
      - '*/*'
      description: Creates a new spending.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: spending to create
        in: body
        name: investment
        required: true
        schema:
          $ref: '#/definitions/finance.CreateSpendingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.createSpendingResponse'
      summary: Create a spending.
      tags:
      - finance
  /meditation:
    get:
      description: Fetch one or multiple meditation sessions.
      parameters:
      - description: Meditation ID
        in: query
        name: id
        type: string
      - description: start time
        in: query
        name: startTime
        type: integer
      - description: end time
        in: query
        name: endTime
        type: integer
      - description: duration start time
        in: query
        name: durationStart
        type: integer
      - description: duration end time
        in: query
        name: durationEnd
        type: integer
      - description: User ID
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meditation.MeditationDB'
            type: array
      summary: Get meditation sessions
      tags:
      - meditation
    post:
      consumes:
      - '*/*'
      description: Creates a new meditation.
      parameters:
      - description: Meditation to create
        in: body
        name: meditation
        required: true
        schema:
          $ref: '#/definitions/meditation.CreateMeditationRequest'
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meditation.createMeditationResponse'
      summary: Create meditation.
      tags:
      - meditation
  /progress:
    get:
      description: fetch progress and level for a user.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/progress.Response'
      summary: Get progress nad level for a user.
      tags:
      - progress
  /settings:
    delete:
      consumes:
      - '*/*'
      description: Delete plugin-settings for a user if plugin is "" delete all settings.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: Plugin name
        in: query
        name: plugin
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Delete plugin-settings of a user.
      tags:
      - settings
    get:
      description: fetch plugin settings for a user.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: Plugin name
        in: query
        name: plugin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/settings.getPluginSettingResponse'
      summary: Get plugin settings for a user.
      tags:
      - settings
    post:
      consumes:
      - '*/*'
      description: Creates settings for a user.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.CreateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/settings.createOnboardingResponse'
      summary: Create onboarding in backend, set settings.
      tags:
      - settings
  /settings/elevator:
    post:
      consumes:
      - '*/*'
      description: Creates settings for a user for the elevator Plugin.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.ElevatorSettings'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Create settings for the elevator plugin.
      tags:
      - settings
    put:
      consumes:
      - '*/*'
      description: Update settings for the elevator Plugin.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.ElevatorSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update settings for the elevator Plugin.
      tags:
      - settings
  /settings/finance:
    post:
      consumes:
      - '*/*'
      description: Creates settings for a user for the finance Plugin.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.FinanceSettings'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Create settings for the finance plugin.
      tags:
      - settings
    put:
      consumes:
      - '*/*'
      description: Update settings for a user for onr Plugin.
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.FinanceSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: updateFinanceSettings settings for the finance plugin.
      tags:
      - settings
  /settings/meditation:
    post:
      consumes:
      - '*/*'
      description: Creates settings for the meditation plugin
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.MeditationSettings'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Create settings for the meditation Plugin.
      tags:
      - settings
    put:
      consumes:
      - '*/*'
      description: Update settings for a user
      parameters:
      - description: User ID
        in: header
        name: userId
        required: true
        type: string
      - description: onboarding to create
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.MeditationSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update settings for the meditation Plugin.
      tags:
      - settings
  /users:
    get:
      consumes:
      - '*/*'
      description: fetch every user available.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserDB'
            type: array
      summary: Get all users.
      tags:
      - users
    post:
      consumes:
      - '*/*'
      description: creates one user.
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.createUserResponse'
      summary: Create one user.
      tags:
      - users
    put:
      consumes:
      - '*/*'
      description: update a user by id.
      parameters:
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.updateUserRequest'
      - description: User ID
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserDB'
      summary: Update a user.
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - '*/*'
      description: delete a user by id with all its progress in all plugins.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete a user.
      tags:
      - users
    get:
      consumes:
      - '*/*'
      description: fetch a user by id.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserDB'
      summary: Get a user.
      tags:
      - users
swagger: "2.0"